requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.10
pydantic==2.7.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-generativeai==0.5.4
pandas==2.2.3
openpyxl==3.1.2
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7